requirements.txt
fastapi
gradio
requests
sentence-transformers
google-generativeai>=0.3.2
uvicorn
pydantic
python-dotenv
psutil
mlflow